+++ sequence_test.go
@@ -30,8 +30,8 @@ func TestSequence_Update(t *testing.T) {
 	)
 
 	seq.Update(0.5)
-	if seq.Value() != float32(0.5) {
-		t.Errorf("expected current to be %v, got %v", float32(0.5), seq.Value())
+	if seq.Value() != float64(0.5) {
+		t.Errorf("expected current to be %v, got %v", float64(0.5), seq.Value())
 	}
 	if seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be false, got %v", seq.IsActiveTweenFinished())
@@ -55,17 +55,17 @@ func TestSequence_Reset(t *testing.T) {
 	if seq.Index() != 0 {
 		t.Errorf("expected index to be %v, got %v", 0, seq.Index())
 	}
-	if seq.Tweens[0].time != float32(0.0) {
-		t.Errorf("expected Tweens[0].time to be %v, got %v", float32(0.0), seq.Tweens[0].time)
+	if seq.Tweens[0].time != float64(0.0) {
+		t.Errorf("expected Tweens[0].time to be %v, got %v", float64(0.0), seq.Tweens[0].time)
 	}
-	if seq.Tweens[0].overflow != float32(0.0) {
-		t.Errorf("expected Tweens[0].overflow to be %v, got %v", float32(0.0), seq.Tweens[0].overflow)
+	if seq.Tweens[0].overflow != float64(0.0) {
+		t.Errorf("expected Tweens[0].overflow to be %v, got %v", float64(0.0), seq.Tweens[0].overflow)
 	}
-	if seq.Tweens[1].time != float32(0.0) {
-		t.Errorf("expected Tweens[1].time to be %v, got %v", float32(0.0), seq.Tweens[1].time)
+	if seq.Tweens[1].time != float64(0.0) {
+		t.Errorf("expected Tweens[1].time to be %v, got %v", float64(0.0), seq.Tweens[1].time)
 	}
-	if seq.Tweens[1].overflow != float32(0.0) {
-		t.Errorf("expected Tweens[1].overflow to be %v, got %v", float32(0.0), seq.Tweens[1].overflow)
+	if seq.Tweens[1].overflow != float64(0.0) {
+		t.Errorf("expected Tweens[1].overflow to be %v, got %v", float64(0.0), seq.Tweens[1].overflow)
 	}
 }
 
@@ -76,8 +76,8 @@ func TestSequence_CompleteFirst(t *testing.T) {
 	)
 
 	seq.Update(1.0)
-	if seq.Value() != float32(1.0) {
-		t.Errorf("expected current to be %v, got %v", float32(1.0), seq.Value())
+	if seq.Value() != float64(1.0) {
+		t.Errorf("expected current to be %v, got %v", float64(1.0), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -97,8 +97,8 @@ func TestSequence_OverflowSecond(t *testing.T) {
 	)
 
 	seq.Update(1.5)
-	if seq.Value() != float32(1.5) {
-		t.Errorf("expected current to be %v, got %v", float32(1.5), seq.Value())
+	if seq.Value() != float64(1.5) {
+		t.Errorf("expected current to be %v, got %v", float64(1.5), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -119,8 +119,8 @@ func TestSequence_OverflowAndComplete(t *testing.T) {
 	)
 
 	seq.Update(3.5)
-	if seq.Value() != float32(3.0) {
-		t.Errorf("expected current to be %v, got %v", float32(3.0), seq.Value())
+	if seq.Value() != float64(3.0) {
+		t.Errorf("expected current to be %v, got %v", float64(3.0), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -141,8 +141,8 @@ func TestSequence_Loops(t *testing.T) {
 	)
 	seq.SetLoop(2)
 	seq.Update(5.25)
-	if seq.Value() != float32(2.25) {
-		t.Errorf("expected current to be %v, got %v", float32(2.25), seq.Value())
+	if seq.Value() != float64(2.25) {
+		t.Errorf("expected current to be %v, got %v", float64(2.25), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -158,8 +158,8 @@ func TestSequence_Loops(t *testing.T) {
 	}
 
 	seq.Update(0.75)
-	if seq.Value() != float32(3) {
-		t.Errorf("expected current to be %v, got %v", float32(3), seq.Value())
+	if seq.Value() != float64(3) {
+		t.Errorf("expected current to be %v, got %v", float64(3), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -183,8 +183,8 @@ func TestSequence_LoopsForever(t *testing.T) {
 	)
 	seq.SetLoop(-1)
 	seq.Update(3*1_000_000 + 2.25)
-	if seq.Value() != float32(2.25) {
-		t.Errorf("expected current to be %v, got %v", float32(2.25), seq.Value())
+	if seq.Value() != float64(2.25) {
+		t.Errorf("expected current to be %v, got %v", float64(2.25), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -209,8 +209,8 @@ func TestSequence_Yoyos(t *testing.T) {
 
 	seq.YoyoEnabled = true
 	seq.Update(5.75)
-	if seq.Value() != float32(0.25) {
-		t.Errorf("expected current to be %v, got %v", float32(0.25), seq.Value())
+	if seq.Value() != float64(0.25) {
+		t.Errorf("expected current to be %v, got %v", float64(0.25), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -226,8 +226,8 @@ func TestSequence_Yoyos(t *testing.T) {
 	}
 
 	seq.Update(0.25)
-	if seq.Value() != float32(0) {
-		t.Errorf("expected current to be %v, got %v", float32(0), seq.Value())
+	if seq.Value() != float64(0) {
+		t.Errorf("expected current to be %v, got %v", float64(0), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -252,8 +252,8 @@ func TestSequence_YoyosAndLoops(t *testing.T) {
 	seq.YoyoEnabled = true
 	seq.SetLoop(2)
 	seq.Update(7.25)
-	if seq.Value() != float32(1.25) {
-		t.Errorf("expected current to be %v, got %v", float32(1.25), seq.Value())
+	if seq.Value() != float64(1.25) {
+		t.Errorf("expected current to be %v, got %v", float64(1.25), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -269,8 +269,8 @@ func TestSequence_YoyosAndLoops(t *testing.T) {
 	}
 
 	seq.Update(4.75)
-	if seq.Value() != float32(0) {
-		t.Errorf("expected current to be %v, got %v", float32(0), seq.Value())
+	if seq.Value() != float64(0) {
+		t.Errorf("expected current to be %v, got %v", float64(0), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -296,8 +296,8 @@ func TestSequence_SetReverse(t *testing.T) {
 
 	// Normal operation
 	seq.Update(2.25)
-	if seq.Value() != float32(2.25) {
-		t.Errorf("expected current to be %v, got %v", float32(2.25), seq.Value())
+	if seq.Value() != float64(2.25) {
+		t.Errorf("expected current to be %v, got %v", float64(2.25), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -316,8 +316,8 @@ func TestSequence_SetReverse(t *testing.T) {
 
 	// Goes in reverse
 	seq.Update(2.0)
-	if seq.Value() != float32(0.25) {
-		t.Errorf("expected current to be %v, got %v", float32(0.25), seq.Value())
+	if seq.Value() != float64(0.25) {
+		t.Errorf("expected current to be %v, got %v", float64(0.25), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -337,8 +337,8 @@ func TestSequence_SetReverse(t *testing.T) {
 
 	// Consumes a loop at the start!, resets to the end, continues in reverse
 	seq.Update(2.0)
-	if seq.Value() != float32(1.25) {
-		t.Errorf("expected current to be %v, got %v", float32(1.25), seq.Value())
+	if seq.Value() != float64(1.25) {
+		t.Errorf("expected current to be %v, got %v", float64(1.25), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -358,8 +358,8 @@ func TestSequence_SetReverse(t *testing.T) {
 
 	// Hits the beginning, no more loops, ends
 	seq.Update(2.0)
-	if seq.Value() != float32(0.0) {
-		t.Errorf("expected current to be %v, got %v", float32(0.0), seq.Value())
+	if seq.Value() != float64(0.0) {
+		t.Errorf("expected current to be %v, got %v", float64(0.0), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -389,8 +389,8 @@ func TestSequence_SetReverseWithYoyo(t *testing.T) {
 
 	// Standard operation
 	seq.Update(2.25)
-	if seq.Value() != float32(2.25) {
-		t.Errorf("expected current to be %v, got %v", float32(2.25), seq.Value())
+	if seq.Value() != float64(2.25) {
+		t.Errorf("expected current to be %v, got %v", float64(2.25), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -409,8 +409,8 @@ func TestSequence_SetReverseWithYoyo(t *testing.T) {
 
 	// Goes in reverse
 	seq.Update(2.0)
-	if seq.Value() != float32(0.25) {
-		t.Errorf("expected current to be %v, got %v", float32(0.25), seq.Value())
+	if seq.Value() != float64(0.25) {
+		t.Errorf("expected current to be %v, got %v", float64(0.25), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -427,8 +427,8 @@ func TestSequence_SetReverseWithYoyo(t *testing.T) {
 
 	// Consumes a loop at the start, despite not reaching the end yet, and continues
 	seq.Update(2.0)
-	if seq.Value() != float32(1.75) {
-		t.Errorf("expected current to be %v, got %v", float32(1.75), seq.Value())
+	if seq.Value() != float64(1.75) {
+		t.Errorf("expected current to be %v, got %v", float64(1.75), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -445,8 +445,8 @@ func TestSequence_SetReverseWithYoyo(t *testing.T) {
 
 	// Hits the end, yoyos
 	seq.Update(2.0)
-	if seq.Value() != float32(2.25) {
-		t.Errorf("expected current to be %v, got %v", float32(2.25), seq.Value())
+	if seq.Value() != float64(2.25) {
+		t.Errorf("expected current to be %v, got %v", float64(2.25), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -468,8 +468,8 @@ func TestSequence_SetReverseWithYoyo(t *testing.T) {
 
 	// Hits the end again, yoyos the same
 	seq.Update(1.5)
-	if seq.Value() != float32(2.25) {
-		t.Errorf("expected current to be %v, got %v", float32(2.25), seq.Value())
+	if seq.Value() != float64(2.25) {
+		t.Errorf("expected current to be %v, got %v", float64(2.25), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -486,8 +486,8 @@ func TestSequence_SetReverseWithYoyo(t *testing.T) {
 
 	// Consumes a loop at the start like normal, no more loops, end
 	seq.Update(2.5)
-	if seq.Value() != float32(0.0) {
-		t.Errorf("expected current to be %v, got %v", float32(0.0), seq.Value())
+	if seq.Value() != float64(0.0) {
+		t.Errorf("expected current to be %v, got %v", float64(0.0), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -513,8 +513,8 @@ func TestSequence_SetReverseAfterComplete(t *testing.T) {
 
 	// Normal operation
 	seq.Update(3.0)
-	if seq.Value() != float32(3.0) {
-		t.Errorf("expected current to be %v, got %v", float32(3.0), seq.Value())
+	if seq.Value() != float64(3.0) {
+		t.Errorf("expected current to be %v, got %v", float64(3.0), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -534,8 +534,8 @@ func TestSequence_SetReverseAfterComplete(t *testing.T) {
 
 	// Goes in reverse
 	seq.Update(2.0)
-	if seq.Value() != float32(1.0) {
-		t.Errorf("expected current to be %v, got %v", float32(1.0), seq.Value())
+	if seq.Value() != float64(1.0) {
+		t.Errorf("expected current to be %v, got %v", float64(1.0), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -570,8 +570,8 @@ func TestSequence_Remove(t *testing.T) {
 		t.Errorf("expected 4 tweens, got %v", len(seq.Tweens))
 	}
 	seq.Update(2.5)
-	if seq.Value() != float32(3.5) {
-		t.Errorf("expected current to be %v, got %v", float32(3.5), seq.Value())
+	if seq.Value() != float64(3.5) {
+		t.Errorf("expected current to be %v, got %v", float64(3.5), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -619,8 +619,8 @@ func TestSequence_Has(t *testing.T) {
 		t.Errorf("expected HasTweens() to be false, got true")
 	}
 	seq.Update(1)
-	if seq.Value() != float32(0) {
-		t.Errorf("expected current to be %v, got %v", float32(0), seq.Value())
+	if seq.Value() != float64(0) {
+		t.Errorf("expected current to be %v, got %v", float64(0), seq.Value())
 	}
 	if seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be false, got %v", seq.IsActiveTweenFinished())
@@ -637,8 +637,8 @@ func TestSequence_SetIndex(t *testing.T) {
 	)
 	seq.SetIndex(1)
 	seq.Update(1.5)
-	if seq.Value() != float32(2) {
-		t.Errorf("expected current to be %v, got %v", float32(2), seq.Value())
+	if seq.Value() != float64(2) {
+		t.Errorf("expected current to be %v, got %v", float64(2), seq.Value())
 	}
 	if !seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be true, got %v", seq.IsActiveTweenFinished())
@@ -670,8 +670,8 @@ func TestSequence_RealWorld(t *testing.T) {
 	}
 
 	seq.Update(1)
-	if seq.Value() != float32(1) {
-		t.Errorf("expected current to be %v, got %v", float32(1), seq.Value())
+	if seq.Value() != float64(1) {
+		t.Errorf("expected current to be %v, got %v", float64(1), seq.Value())
 	}
 	if seq.IsActiveTweenFinished() {
 		t.Errorf("expected finishedTween to be false, got %v", seq.IsActiveTweenFinished())
@@ -681,8 +681,8 @@ func TestSequence_RealWorld(t *testing.T) {
 	}
 
 	seq.Update(1)
-	if seq.Value() != float32(2) {
-		t.Errorf("expected current to be %v, got %v", float32(2), seq.Value())
+	if seq.Value() != float64(2) {
+		t.Errorf("expected current to be %v, got %v", float64(2), seq.Value())
 	}
 	if seq.Index() != 1 {
 		t.Errorf("expected index to be %v, got %v", 1, seq.Index())
